made onLogin an event like any other
