+++ src/engine/events.js
@@ -1,10 +1,13 @@
 const { AppConfig } = require("../../app.config");
-const { Message } = require('discord.js');
+const { Message, Client } = require('discord.js');
+const { LiteralConstants } = require("../models/literal.constants");
+const { Logger } = require('../utils/logger');
 
 /**
  * The callback for an event.
  *
  * @callback EventCallback
+ * @param {Client} client
  * @param {Message} message
  * @param {ErrorCallback} onError
  */
@@ -25,10 +28,33 @@ const { Message } = require('discord.js');
 function events(appConfig) { 
     const discordHelpers = appConfig.DISCORD_HELPERS;
         return new Map([
-        ["guildDelete", async (input, onError) => {
+        ["ready", async (client, input, onError) => {
+            for(let guild of client.guilds.cache.array()){
+                if(guild && guild.me){
+                    guild.me.setNickname(LiteralConstants.BOT_NAME_OFFLINE);
+                    // TODO: notify servers of changes since last login
+                }
+                // logger.log(`${guild.name} | ${guild.id}`);
+                new Logger(appConfig.DISCORD_HELPERS.getGuildId(guild)).log(LiteralConstants.LOG_SESSION_START);
+                const dummyMessage = {
+                    guild: guild,
+                    channel:{send(){}}
+                }
+                try{
+                    const listening = appConfig.CONFIG_STORAGE.getProperty(guild, 'listening');
+                    if(listening == true){
+                        await appConfig.COMMANDS(appConfig).find(c => c.aliases.includes('start')).callback(dummyMessage);
+                    }
+                }catch(e){
+                    console.log(e);
+                    onError(dummyMessage, `Error staring listener for Guild: ${appConfig.DISCORD_HELPERS.getGuildId(guild)}: ${e.stack ? e.stack : e}`);
+                }
+            }
+        }],
+        ["guildDelete", async (client, input, onError) => {
             return appConfig.CONFIG_STORAGE.deleteGuildConfig(input);
         }],
-        ["message", async (message, onError) => { 
+        ["message", async (client, message, onError) => { 
             if(discordHelpers.isDm(message) || discordHelpers.isBot(message)){
                 return;
             }
